+++ checker/src/main/java/org/checkerframework/checker/index/lowerbound/LowerBoundAnnotatedTypeFactory.java
@@ -167,7 +167,7 @@ public class LowerBoundAnnotatedTypeFactory extends BaseAnnotatedTypeFactoryForI
       }
       return;
     }
-    if (qualHierarchy.isSubtype(anm, type.getAnnotationInHierarchy(UNKNOWN))) {
+    if (typeHierarchy.isSubtypeShallowEffective(anm, type)) {
       type.replaceAnnotation(anm);
     }
   }
@@ -354,8 +354,9 @@ public class LowerBoundAnnotatedTypeFactory extends BaseAnnotatedTypeFactoryForI
         AnnotatedTypeMirror leftType = getAnnotatedType(left);
         AnnotatedTypeMirror rightType = getAnnotatedType(right);
         type.replaceAnnotation(
-            qualHierarchy.greatestLowerBound(
-                leftType.getAnnotationInHierarchy(POS), rightType.getAnnotationInHierarchy(POS)));
+            qualHierarchy.greatestLowerBoundShallow(
+                leftType.getAnnotationInHierarchy(POS), leftType.getUnderlyingType(),
+                rightType.getAnnotationInHierarchy(POS), rightType.getUnderlyingType()));
       }
       return super.visitMethodInvocation(tree, type);
     }