+++ framework/src/main/java/org/checkerframework/common/wholeprograminference/WholeProgramInferenceScenesStorage.java
@@ -590,7 +590,10 @@ public class WholeProgramInferenceScenesStorage
       // If the inferred type is a subtype of the upper bounds of the
       // current type on the source code, halt.
       if (upperAnnos.size() == rhsATM.getAnnotations().size()
-          && atypeFactory.getQualifierHierarchy().isSubtype(rhsATM.getAnnotations(), upperAnnos)) {
+          && atypeFactory
+              .getQualifierHierarchy()
+              .isSubtypeShallow(
+                  rhsATM.getAnnotations(), rhsTM, upperAnnos, lhsATM.getUnderlyingType())) {
         return;
       }
     }
@@ -648,7 +651,14 @@ public class WholeProgramInferenceScenesStorage
       // amJaif only contains annotations from the jaif, so it might be missing
       // an annotation in the hierarchy
       if (amJaif != null) {
-        amSource = atypeFactory.getQualifierHierarchy().leastUpperBound(amSource, amJaif);
+        amSource =
+            atypeFactory
+                .getQualifierHierarchy()
+                .leastUpperBoundShallow(
+                    amSource,
+                    sourceCodeATM.getUnderlyingType(),
+                    amJaif,
+                    jaifATM.getUnderlyingType());
       }
       annosToReplace.add(amSource);
     }