+++ framework/src/main/java/org/checkerframework/framework/type/GenericAnnotatedTypeFactory.java
@@ -2573,6 +2573,11 @@ public abstract class GenericAnnotatedTypeFactory<
         }
         return false;
 
+      case EXECUTABLE:
+      case MODULE:
+      case PACKAGE:
+        return false;
+
       default:
         throw new BugInCF("isRelevantHelper(%s): Unexpected TypeKind %s", tm, tm.getKind());
     }